feat: extract search form into reusable LiveComponent and fix FOUC
- Extract search form markup and event handling into SearchFormComponent
- Replace duplicate form implementations with reusable component instances
- Fix Flash of Unstyled Content (FOUC) issues by removing CSS transitions:
  - Remove transition-colors from Sign In/Sign Out links to prevent text color flicker
  - Remove transition-colors from Learn More button to fix background/text color flash
  - Remove transition-all from search input to prevent resize flicker on initial render
  - Replace with transition-colors only for search input border states
